Make 32 bit support optional
It would be useful to make it eligible if you want to have 32-bit support or not. Many of these useful 32-bit libraries for Steam, Wine or Proton... can be isolated with the Flatpak sandbox.
On the other hand, installing fewer packages at the OS level reduces the attack surface.